# fluxcd-trainer-dir/podinfo/pkg/api/http/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Go microservice template for Kubernetes.",
        "title": "Podinfo API",
        "contact": {
            "name": "Source Code",
            "url": "https://github.com/stefanprodan/podinfo"
        },
        "license": {
            "name": "MIT License",
            "url": "https://github.com/stefanprodan/podinfo/blob/master/LICENSE"
        },
        "version": "2.0"
    },
    "basePath": "/",
    "paths": {
...
        "/chunked/{seconds}": {
            "get": {
                "description": "uses transfer-encoding type chunked to give a partial response and then waits for the specified period",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "HTTP API"
                ],
                "summary": "Chunked transfer encoding",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "seconds to wait for",
                        "name": "seconds",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.MapResponse"
                        }
                    }
                }
            }
        },
        "/delay/{seconds}": {
            "get": {
                "description": "waits for the specified period",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "HTTP API"
                ],
                "summary": "Delay",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "seconds to wait for",
                        "name": "seconds",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.MapResponse"
                        }
                    }
                }
            }
        },
...
        "/store/{hash}": {
            "get": {
                "description": "returns the content of the file /data/hash if exists",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "HTTP API"
                ],
                "summary": "Download file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "hash value",
                        "name": "hash",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "file",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "http.MapResponse": {
            "type": "object",
            "additionalProperties": {
                "type": "string"
            }
        },
...
    }
}